[SPARK-47172] Addressing comments & Refactoring AuthEngine test suite into two classes
